refactor(config): move resolved/authored types out of common.rs
common.rs is now shared serde primitives only (config enums, ConfigName,
webhook route shapes, serde defaults). The *Resolved component shapes and
DeploymentResolved move to resolve.rs (the stage that produces them); the
authored WASM structs and cron config move to authored.rs. The webhook and
cron submodules are dissolved into top-level items in their owning files, so
mod.rs no longer needs their explicit re-exports.
