+++ src/config/toml/mod.rs
@@ -80,12 +80,6 @@ pub(crate) use common::*;
 pub(crate) use resolve::*;
 pub(crate) use server::*;
 
-pub(crate) use crate::config::toml::common::cron::CronComponentConfigToml;
-pub(crate) use crate::config::toml::common::webhook::{
-    WebhookJsComponentConfigResolved, WebhookRoute, WebhookRouteDetail,
-    WebhookWasmComponentConfigResolved, default_external_server_name,
-};
-
 const DEFAULT_SQLITE_DIR_IF_PROJECT_DIRS: &str =
     const_format::formatcp!("{}obelisk-sqlite", DATA_DIR_PREFIX);
 const DEFAULT_SQLITE_DIR: &str = "obelisk-sqlite";
